feat: change assets by component
